rust/point-mutations: i32 -> usize (tests do not specify type)

+++ rust/point-mutations/src/lib.rs
@@ -1,10 +1,10 @@
-pub fn hamming_distance(a: &str, b: &str) -> Result<i32, &'static str> {
+pub fn hamming_distance(a: &str, b: &str) -> Result<usize, &'static str> {
     match a.len() == b.len() {
         true => {
             Ok(a.chars()
                 .zip(b.chars())
                 .filter(|x| x.0 != x.1)
-                .count() as i32)
+                .count())
         }
         false => Err("inputs of different length"),
     }
